Allow charms to not have haproxy admin socket enabled
This commit refactors the code a little to place the logic for
draining the haproxy backend servers inside a Unit/APIUnit class.
As part of this, the unit will determine whether or not the haproxy
admin socket is enabled on the unit itself and will either perform
the drain or not.